Unify install script for backend and server

+++ scripts/install_backend.sh
@@ -20,42 +20,43 @@ fi
 SCRIPT=`realpath $0`
 SCRIPTDIR=`dirname $SCRIPT`
 
-# In case of update remove first old jupyterhub version
-sudo rm -rf /opt/jupyterhub
+if [ $WODTYPE = "backend" ]; then
+	# In case of update remove first old jupyterhub version
+	sudo rm -rf /opt/jupyterhub
+fi
 
 cat > $SCRIPTDIR/wod.sh << EOF
 # This main dir is computed
 export JUPPROC=`dirname $SCRIPTDIR`
 EOF
 cat >> $SCRIPTDIR/wod.sh << 'EOF'
-# These 2 dirs have fixed names by default that you can change in this file
+# These 3 dirs have fixed names by default that you can change in this file
 # they are placed as sister dirs wrt JUPPROC
 PJUPPROC=`dirname $JUPPROC`
 export JUPPRIV=$PJUPPROC/wod-private
-export JUPNOBO=$PJUPPROC/wod-notebooks
-# This dir is also fixed by default and can be changed as needed
-export STUDDIR=/student
+export SRVDIR=$PJUPPROC/wod-server
 WODANSOPT=""
 # Manages private inventory if any
 if [ -f $JUPPRIV/ansible/inventory ]; then
 	WODANSOPT="-i $JUPPRIV/ansible/inventory"
-export WODANSOPT
+	export WODANSOPT
 fi
 EOF
+if [ $WODTYPE = "backend" ]; then
+	cat >> $SCRIPTDIR/wod.sh << 'EOF'
+# This dir is also fixed by default and can be changed as needed
+export JUPNOBO=$PJUPPROC/wod-notebooks
+export STUDDIR=/student
+EOF
+fi
+
 chmod 755 $SCRIPTDIR/wod.sh
 source $SCRIPTDIR/wod.sh
 
-if ! command -v ansible-galaxy &> /dev/null
-then
-    echo "ansible-galaxy could not be found, please install ansible"
-    exit -1
-fi
-
-cd $SCRIPTDIR/../ansible
 SHORTNAME="`hostname -s`"
 FULLNAME=`ansible-inventory -i inventory --list | jq -r '._meta.hostvars | to_entries[] | .key' | grep -E "^$SHORTNAME(\.|$)"`
 if [ _"$FULLNAME" = _"" ]; then
-        echo "This machine is not a jupyterhub machine, defined in the ansible inventory so can't be installed"
+        echo "This machine is not a $WODTYPE machine, defined in the ansible inventory so can't be installed"
         exit -1
 fi
 PBKDIR=`ansible-inventory -i inventory --list | jq -r "._meta.hostvars | to_entries[] | select(.key == \"$FULLNAME\") | .value.PBKDIR"`
@@ -64,14 +65,22 @@ WODDISTRIB=`grep -E '^ID=' /etc/os-release | cut -d= -f2 | sed 's/"//g'`-`grep -
 # Another way using ansible
 #DISTRIB=`ansible -m gather_facts -i inventory $FULLNAME | perl -p -e "s/$FULLNAME \| SUCCESS => //" | jq -r ".ansible_facts | .ansible_distribution"`
 #DVER=`ansible -m gather_facts -i inventory $FULLNAME | perl -p -e "s/$FULLNAME \| SUCCESS => //" | jq -r ".ansible_facts | .ansible_distribution_major_version"`
-if [ $WODDISTRIB = "centos-7" ] || [ $WODDISTRIB = "ubuntu-20.04" ]; then
-	# Older distributions require an older version of the collection to work.
-	# See https://github.com/ansible-collections/community.general
-	ansible-galaxy collection install --force-with-deps community.general:4.8.5
-else
-	ansible-galaxy collection install community.general
+if [ $WODTYPE = "backend" ]; then
+	if ! command -v ansible-galaxy &> /dev/null
+	then
+	    echo "ansible-galaxy could not be found, please install ansible"
+	    exit -1
+	fi
+	if [ $WODDISTRIB = "centos-7" ] || [ $WODDISTRIB = "ubuntu-20.04" ]; then
+		# Older distributions require an older version of the collection to work.
+		# See https://github.com/ansible-collections/community.general
+		ansible-galaxy collection install --force-with-deps community.general:4.8.5
+	else
+		ansible-galaxy collection install community.general
+	fi
+	ansible-galaxy collection install ansible.posix
 fi
-ansible-galaxy collection install ansible.posix
+
 
 SCRIPTREL=`echo $SCRIPT | perl -p -e "s|$JUPPROC||"`
 if [ -x $JUPPRIV/$SCRIPTREL ];
@@ -80,14 +89,19 @@ then
 	$JUPPRIV/$SCRIPTREL
 fi
 
-# Automatic Installation script for jupyterhub 
+cd $SCRIPTDIR/../ansible
+# Automatic Installation script for the system 
 ansible-playbook -i inventory --limit $PBKDIR -e "LDAPSETUP=0 -e APPMIN=0 -e APPMAX=0" install_backend.yml
 ansible-playbook -i inventory --limit $PBKDIR check_backend.yml
 
-if [ -f $JUPPRIV/ansible/install_backend.yml ]; then
-	ansible-playbook -i inventory $WODANSOPT --limit $PBKDIR -e "LDAPSETUP=0 -e APPMIN=0 -e APPMAX=0" install_backend.yml
+ANSPLAYOPT=""
+if [ $WODTYPE = "backend" ]; then
+	ANSPLAYOPT="-e "LDAPSETUP=0 -e APPMIN=0 -e APPMAX=0"
 fi
-if [ -f $JUPPRIV/ansible/check_backend.yml ]; then
-	ansible-playbook -i inventory $WODANSOPT --limit $PBKDIR check_backend.yml
+if [ -f $JUPPRIV/ansible/install_$WODTYPE.yml ]; then
+	ansible-playbook -i inventory $WODANSOPT --limit $PBKDIR $ANSPLAYOPT install_$WODTYPE.yml
+fi
+if [ -f $JUPPRIV/ansible/check_$WODTYPE.yml ]; then
+	ansible-playbook -i inventory $WODANSOPT --limit $PBKDIR check_$WODTYPE.yml
 fi
 date
